moving Clock class to it's own header and source file
